Hotfix: Prevent FileNotFoundException on Mac/Linux
This plugin keeps crashing at random times because of a hardcoded "\\" file separator.
This is because every OS other than Windows uses `/` as a path separator and I'm trying
to use this amazing plugin from one of them  :¬)

This pull request is for my fellow Machintosh, BSD and Linux users.